pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Joe Nudell <jnudell@hks.harvard.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.2"}
sqlalchemy = "^2.0.36"
aiosqlite = "^0.20.0"
python-glowplug = "^0.3.10"
pydantic = "^2.9.2"
greenlet = "^3.1.1"
openai = "^1.52.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
